Render audio samples as double instead of int32.
- Recompute loudness for related test vectors.

PiperOrigin-RevId: 669410759

+++ iamf/cli/renderer/audio_element_renderer_base.cc
@@ -23,7 +23,7 @@ namespace iamf_tools {
 AudioElementRendererBase::~AudioElementRendererBase() {}
 
 absl::Status AudioElementRendererBase::Flush(
-    std::vector<int32_t>& rendered_samples) {
+    std::vector<double>& rendered_samples) {
   absl::MutexLock lock(&mutex_);
   rendered_samples.insert(rendered_samples.end(), rendered_samples_.begin(),
                           rendered_samples_.end());
